@@ -1,4 +1,5 @@
-﻿using DesafioFundamentos.Models;
+﻿using System.Security.Cryptography;
+using DesafioFundamentos.Models;
 
 // Coloca o encoding para UTF8 para exibir acentuação
 Console.OutputEncoding = System.Text.Encoding.UTF8;
@@ -6,15 +7,14 @@ Console.OutputEncoding = System.Text.Encoding.UTF8;
 decimal precoInicial = 0;
 decimal precoPorHora = 0;
 
-Console.WriteLine("Seja bem vindo ao sistema de estacionamento!\n" +
-                  "Digite o preço inicial:");
-precoInicial = Convert.ToDecimal(Console.ReadLine());
+Console.WriteLine("Seja bem vindo ao sistema de estacionamento!\n" + "Digite o preço inicial:");
+_ = decimal.TryParse(Console.ReadLine().Replace(".", ","), out precoInicial);
 
 Console.WriteLine("Agora digite o preço por hora:");
-precoPorHora = Convert.ToDecimal(Console.ReadLine());
+_ = decimal.TryParse(Console.ReadLine().Replace(".", ","), out precoPorHora);
 
 // Instancia a classe Estacionamento, já com os valores obtidos anteriormente
-Estacionamento es = new Estacionamento(precoInicial, precoPorHora);
+Estacionamento es = new(precoInicial, precoPorHora);
 
 string opcao = string.Empty;
 bool exibirMenu = true;
@@ -22,38 +22,38 @@ bool exibirMenu = true;
 // Realiza o loop do menu
 while (exibirMenu)
 {
-    Console.Clear();
-    Console.WriteLine("Digite a sua opção:");
-    Console.WriteLine("1 - Cadastrar veículo");
-    Console.WriteLine("2 - Remover veículo");
-    Console.WriteLine("3 - Listar veículos");
-    Console.WriteLine("4 - Encerrar");
+	Console.Clear();
+	Console.WriteLine("Digite a sua opção:");
+	Console.WriteLine("1 - Cadastrar veículo");
+	Console.WriteLine("2 - Remover veículo");
+	Console.WriteLine("3 - Listar veículos");
+	Console.WriteLine("4 - Encerrar");
 
-    switch (Console.ReadLine())
-    {
-        case "1":
-            es.AdicionarVeiculo();
-            break;
+	switch (Console.ReadLine())
+	{
+		case "1":
+			es.AdicionarVeiculo();
+			break;
 
-        case "2":
-            es.RemoverVeiculo();
-            break;
+		case "2":
+			es.RemoverVeiculo();
+			break;
 
-        case "3":
-            es.ListarVeiculos();
-            break;
+		case "3":
+			es.ListarVeiculos();
+			break;
 
-        case "4":
-            exibirMenu = false;
-            break;
+		case "4":
+			exibirMenu = false;
+			break;
 
-        default:
-            Console.WriteLine("Opção inválida");
-            break;
-    }
+		default:
+			Console.WriteLine("Opção inválida");
+			break;
+	}
 
-    Console.WriteLine("Pressione uma tecla para continuar");
-    Console.ReadLine();
+	Console.WriteLine("\nPressione uma tecla para continuar");
+	Console.ReadLine();
 }
 
 Console.WriteLine("O programa se encerrou");
